refactor(backend): split connector lifecycle modules
