+++ backend/internal/api/connector_target_handlers.go
@@ -9,7 +9,6 @@ import (
 	"net/http"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/aipermission/aipermission/backend/internal/connectorapi"
 	"github.com/aipermission/aipermission/backend/internal/connectors"
@@ -856,275 +855,6 @@ func (s connectorTargetHandlers) finalizeDeletedConnectorTarget(w http.ResponseW
 	return true
 }
 
-func (s connectorTargetHandlers) listConnectorCredentialProfiles(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	if _, err := store.GetTarget(r.Context(), targetID); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	profiles, err := store.ListCredentialProfiles(r.Context(), targetID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"items": profileSummaries(profiles)})
-}
-
-func (s connectorTargetHandlers) createConnectorCredentialProfile(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	var request createConnectorCredentialProfileRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	target, err := store.GetTarget(r.Context(), targetID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(target.ConnectorKind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	preparedProfile, ok := s.prepareConnectorCredentialProfileInput(w, r, runtime, connector, createProfileAdapterRequest(request), true, nil, "")
-	if !ok {
-		return
-	}
-	tx, err := runtime.database.BeginTx(r.Context(), nil)
-	if err != nil {
-		writeInternalError(w)
-		return
-	}
-	defer tx.Rollback()
-	txStore := connectortargets.NewTxStore(tx)
-	if adapter := connectorCredentialProfileLifecycleAdapterFor(target.ConnectorKind); adapter != nil {
-		if err := adapter.BeforeCreateCredentialProfile(r.Context(), runtime, txStore, target); err != nil {
-			handleConnectorTargetError(w, err)
-			return
-		}
-	}
-	profile, err := txStore.CreateCredentialProfile(r.Context(), connectortargets.CreateCredentialProfileInput{
-		TargetID:            target.ID,
-		ConnectorKind:       target.ConnectorKind,
-		Kind:                preparedProfile.Kind,
-		Label:               preparedProfile.Label,
-		Public:              preparedProfile.Public,
-		EncryptedSecretJSON: preparedProfile.EncryptedSecretJSON,
-		RiskLabel:           preparedProfile.RiskLabel,
-	})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if err := s.ensureConnectorRuntimeSurfacesForProfile(r.Context(), txStore, target, profile); err != nil {
-		writeInternalError(w)
-		return
-	}
-	if err := tx.Commit(); err != nil {
-		writeInternalError(w)
-		return
-	}
-	s.writeAudit(r.Context(), runtime, "user", nil, 0, "connector.profile.created", map[string]any{
-		"target_id":      target.ID,
-		"profile_id":     profile.ID,
-		"connector_kind": target.ConnectorKind,
-		"kind":           profile.Kind,
-		"label":          profile.Label,
-	})
-	writeJSON(w, http.StatusCreated, profileToSummary(profile))
-}
-
-func (s connectorTargetHandlers) updateConnectorCredentialProfile(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	profileID, ok := parsePathInt64(w, r, "profile_id", "profile_id")
-	if !ok {
-		return
-	}
-	var request updateConnectorCredentialProfileRequest
-	if err := decodeJSON(w, r, &request); err != nil {
-		writeError(w, http.StatusBadRequest, "invalid json body")
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	target, err := store.GetTarget(r.Context(), targetID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(target.ConnectorKind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	existingProfile, err := store.GetCredentialProfile(r.Context(), targetID, profileID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if lifecycle, ok := connector.(connectors.ProvisionedCredentialLifecycle); ok {
-		mergedPublic, err := lifecycle.PreserveProvisionedCredentialPublic(connectortargets.CredentialProfileView(existingProfile), request.Public)
-		if err != nil {
-			handleConnectorTargetError(w, connectortargets.ValidationError(err.Error()))
-			return
-		}
-		request.Public = mergedPublic
-	}
-	existingProfileView := connectortargets.CredentialProfileView(existingProfile)
-	preparedProfile, ok := s.prepareConnectorCredentialProfileInput(w, r, runtime, connector, updateProfileAdapterRequest(request), request.Secret != nil, &existingProfileView, existingProfile.EncryptedSecretJSON)
-	if !ok {
-		return
-	}
-	release, err := runtime.vaultDelivery.acquire(r.Context())
-	if err != nil {
-		writeError(w, http.StatusRequestTimeout, "connector credential profile update was canceled")
-		return
-	}
-	defer release()
-	tx, err := runtime.database.BeginTx(r.Context(), nil)
-	if err != nil {
-		writeInternalError(w)
-		return
-	}
-	defer tx.Rollback()
-	txStore := connectortargets.NewTxStore(tx)
-	profile, err := txStore.UpdateCredentialProfile(r.Context(), connectortargets.UpdateCredentialProfileInput{
-		TargetID:            target.ID,
-		ProfileID:           profileID,
-		ConnectorKind:       target.ConnectorKind,
-		Kind:                preparedProfile.Kind,
-		Label:               preparedProfile.Label,
-		Public:              preparedProfile.Public,
-		EncryptedSecretJSON: preparedProfile.EncryptedSecretPtr,
-		RiskLabel:           preparedProfile.RiskLabel,
-	})
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if err := s.ensureConnectorRuntimeSurfacesForProfile(r.Context(), txStore, target, profile); err != nil {
-		writeInternalError(w)
-		return
-	}
-	if err := tx.Commit(); err != nil {
-		writeInternalError(w)
-		return
-	}
-	if err := invalidateVaultSessionsForTargetProfile(
-		r.Context(), runtime, target.ID, profile.ID,
-		"connector credential profile changed; send a fresh Vault request",
-	); err != nil {
-		writeInternalError(w)
-		return
-	}
-	staleRequests, err := s.staleConnectorActionRequestsForTarget(r.Context(), runtime, target.ID, profile.ID, "connector credential profile was updated; ask the AI to send a fresh request", false)
-	if err != nil {
-		writeInternalError(w)
-		return
-	}
-	s.writeAudit(r.Context(), runtime, "user", nil, 0, "connector.profile.updated", map[string]any{
-		"target_id":                target.ID,
-		"profile_id":               profile.ID,
-		"connector_kind":           target.ConnectorKind,
-		"kind":                     profile.Kind,
-		"label":                    profile.Label,
-		"stale_connector_requests": staleRequests,
-	})
-	writeJSON(w, http.StatusOK, profileToSummary(profile))
-}
-
-func (s connectorTargetHandlers) deleteConnectorCredentialProfile(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	profileID, ok := parsePathInt64(w, r, "profile_id", "profile_id")
-	if !ok {
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	target, err := store.GetTarget(r.Context(), targetID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	profile, err := store.GetCredentialProfile(r.Context(), targetID, profileID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	release, err := runtime.vaultDelivery.acquire(r.Context())
-	if err != nil {
-		writeError(w, http.StatusRequestTimeout, "connector credential profile deletion was canceled")
-		return
-	}
-	defer release()
-	if err := s.cleanupProvisionedCredentialProfileIfNeeded(r.Context(), runtime, target, profile); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if adapter := connectorCredentialProfileLifecycleAdapterFor(target.ConnectorKind); adapter != nil {
-		if err := adapter.BeforeDeleteCredentialProfile(r.Context(), s, runtime, store, target, profile); err != nil {
-			handleConnectorTargetError(w, err)
-			return
-		}
-	}
-	if err := store.DeleteCredentialProfile(r.Context(), targetID, profileID); err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if err := invalidateVaultSessionsForTargetProfile(
-		r.Context(), runtime, targetID, profileID,
-		"connector credential profile was deleted; send a fresh Vault request",
-	); err != nil {
-		writeInternalError(w)
-		return
-	}
-	staleRequests, err := s.staleConnectorActionRequestsForTarget(r.Context(), runtime, targetID, profileID, "connector credential profile was deleted; ask the AI to send a fresh request", true)
-	if err != nil {
-		writeInternalError(w)
-		return
-	}
-	s.writeAudit(r.Context(), runtime, "user", nil, 0, "connector.profile.deleted", map[string]any{
-		"target_id":                target.ID,
-		"profile_id":               profile.ID,
-		"connector_kind":           target.ConnectorKind,
-		"kind":                     profile.Kind,
-		"label":                    profile.Label,
-		"stale_connector_requests": staleRequests,
-	})
-	w.WriteHeader(http.StatusNoContent)
-}
-
 func (s connectorTargetHandlers) staleConnectorActionRequestsForTarget(ctx context.Context, runtime *databaseRuntime, targetID int64, profileID int64, reason string, includeRunning bool) (int64, error) {
 	if runtime == nil || runtime.database == nil || targetID < 1 {
 		return 0, nil
@@ -1175,138 +905,6 @@ func connectorLifecycleApprovalDrift(profileID int64) string {
 		return "profile"
 	}
 	return "target"
-}
-
-func (s connectorTargetHandlers) testConnectorCredentialProfile(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	profileID, ok := parsePathInt64(w, r, "profile_id", "profile_id")
-	if !ok {
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	loadedTarget, err := store.GetTarget(r.Context(), targetID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	target, profile, err := store.ResolveConnectorActionTarget(r.Context(), connectortargets.ConnectorTargetRef(loadedTarget.ConnectorKind, targetID, profileID))
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	if adapter := connectorCredentialProfileTesterFor(target.ConnectorKind); adapter != nil {
-		adapter.TestCredentialProfile(s, w, r, runtime, target, profile)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(target.ConnectorKind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	testable, ok := connector.(connectors.TestableConnector)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "connector does not support connection tests")
-		return
-	}
-	fullProfile, err := store.GetCredentialProfile(r.Context(), target.ID, profile.ID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	secrets := map[string]any{}
-	if fullProfile.EncryptedSecretJSON != "" {
-		if err := runtime.vault.DecryptJSON(fullProfile.EncryptedSecretJSON, &secrets); err != nil {
-			writeInternalError(w)
-			return
-		}
-	}
-	ctx, cancel := context.WithTimeout(r.Context(), 20*time.Second)
-	defer cancel()
-	start := time.Now()
-	result, err := testable.TestConnection(ctx, connectors.RuntimeContext{
-		Target:       target,
-		Profile:      profile,
-		Secrets:      connectorSecretAccessor{values: secrets},
-		Capabilities: connectorRuntimeCapabilitiesFor(target.ConnectorKind, s.Server, runtime),
-		Events:       noopConnectorEventSink{},
-	})
-	if err != nil {
-		writeJSON(w, http.StatusOK, connectorTargetTestResponse{
-			TargetID:      target.ID,
-			ProfileID:     profile.ID,
-			ConnectorKind: target.ConnectorKind,
-			OK:            false,
-			Status:        string(connectors.TestUnknownError),
-			Message:       s.redactForPersistence(r.Context(), runtime, err.Error()),
-			DurationMS:    time.Since(start).Milliseconds(),
-		})
-		return
-	}
-	writeJSON(w, http.StatusOK, connectorTargetTestResponse{
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ConnectorKind: target.ConnectorKind,
-		OK:            result.Status == connectors.TestOK,
-		Status:        string(result.Status),
-		Message:       s.redactForPersistence(r.Context(), runtime, result.Message),
-		Details:       redactedMapValue(s.redactedConnectorValue(r.Context(), runtime, result.Details, connectorSensitiveOutputFields())),
-		DurationMS:    time.Since(start).Milliseconds(),
-	})
-}
-
-func redactedMapValue(value any) map[string]any {
-	if value == nil {
-		return nil
-	}
-	if typed, ok := value.(map[string]any); ok {
-		return typed
-	}
-	return map[string]any{"value": value}
-}
-
-func (s connectorTargetHandlers) listConnectorCredentialProfileActions(w http.ResponseWriter, r *http.Request) {
-	runtime, ok := s.activeRuntimeOrLocked(w)
-	if !ok {
-		return
-	}
-	targetID, ok := parseID(w, r)
-	if !ok {
-		return
-	}
-	profileID, ok := parsePathInt64(w, r, "profile_id", "profile_id")
-	if !ok {
-		return
-	}
-	store := connectortargets.NewStore(runtime.database)
-	target, profile, err := connectorTargetProfileViews(r.Context(), store, targetID, profileID)
-	if err != nil {
-		handleConnectorTargetError(w, err)
-		return
-	}
-	registry := runtime.connectorRegistry()
-	connector, ok := registry.Get(target.ConnectorKind)
-	if !ok {
-		writeError(w, http.StatusBadRequest, "unsupported connector kind")
-		return
-	}
-	actions, err := connector.GetActionList(r.Context(), target, profile)
-	if err != nil {
-		writeInternalError(w)
-		return
-	}
-	if err := connectors.ValidateActionDefinitions(actions, target.ConnectorKind+" actions"); err != nil {
-		writeInternalError(w)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"items": actions})
 }
 
 func (s connectorTargetHandlers) runConnectorTargetOperation(w http.ResponseWriter, r *http.Request) {
