+++ backend/internal/connectors/postgres/postgres.go
@@ -2,18 +2,10 @@
 package postgresconnector
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"net"
-	"net/url"
-	"os"
-	"os/exec"
 	"regexp"
 	"strconv"
 	"strings"
@@ -484,805 +476,6 @@ func (o queryOutput) ToMap() map[string]any {
 
 func (o queryOutput) DisplayText() string {
 	return o.Result.DisplayText()
-}
-
-type provisionScope struct {
-	AllSchemas bool
-	Schemas    []provisionSchemaScope
-}
-
-type provisionSchemaScope struct {
-	Schema    string
-	AllTables bool
-	Tables    []provisionTableScope
-}
-
-type provisionTableScope struct {
-	Table      string
-	AllColumns bool
-	Columns    []string
-}
-
-func (Connector) ProvisionCredentialProfile(ctx context.Context, runtime connectors.RuntimeContext, input map[string]any) (connectors.ProvisionedCredentialProfile, error) {
-	if runtime.Target.ConnectorKind != Kind {
-		return connectors.ProvisionedCredentialProfile{}, fmt.Errorf("target connector kind must be %s", Kind)
-	}
-	roleName := cleanSimpleIdentifierInput(input, "role_name")
-	if roleName == "" {
-		return connectors.ProvisionedCredentialProfile{}, fmt.Errorf("role_name is required and must be a simple identifier")
-	}
-	profileLabel := strings.TrimSpace(stringInput(input, "profile_label"))
-	if profileLabel == "" {
-		profileLabel = roleName
-	}
-	preset := strings.TrimSpace(stringInput(input, "preset"))
-	switch preset {
-	case "", "read_only":
-		preset = "read_only"
-	case "read_write":
-	default:
-		return connectors.ProvisionedCredentialProfile{}, fmt.Errorf("unsupported preset %q", preset)
-	}
-	scope, err := provisionScopeInput(input)
-	if err != nil {
-		return connectors.ProvisionedCredentialProfile{}, err
-	}
-	password, err := randomCredentialPassword()
-	if err != nil {
-		return connectors.ProvisionedCredentialProfile{}, err
-	}
-	statements, summary, err := provisionRoleStatements(runtime.Target, roleName, password, preset, scope)
-	if err != nil {
-		return connectors.ProvisionedCredentialProfile{}, err
-	}
-
-	ctx, cancel := context.WithTimeout(ctx, queryTimeout)
-	defer cancel()
-	conn, err := connectPostgres(ctx, runtime, false)
-	if err != nil {
-		return connectors.ProvisionedCredentialProfile{}, err
-	}
-	defer conn.Close(context.Background())
-	tx, err := conn.Begin(ctx)
-	if err != nil {
-		return connectors.ProvisionedCredentialProfile{}, fmt.Errorf("start postgres role provisioning transaction: %w", err)
-	}
-	defer tx.Rollback(context.Background())
-	for _, statement := range statements {
-		if _, err := tx.Exec(ctx, statement); err != nil {
-			return connectors.ProvisionedCredentialProfile{}, fmt.Errorf("provision postgres role: %w", err)
-		}
-	}
-	if err := tx.Commit(ctx); err != nil {
-		return connectors.ProvisionedCredentialProfile{}, fmt.Errorf("commit postgres role provisioning: %w", err)
-	}
-
-	public := map[string]any{
-		"username":                  roleName,
-		"managed_by_aipermission":   true,
-		"managed_role_name":         roleName,
-		"managed_admin_profile_id":  runtime.Profile.ID,
-		"managed_admin_profile_ref": runtime.Profile.Label,
-		"managed_preset":            preset,
-		"managed_scope":             summary,
-	}
-	return connectors.ProvisionedCredentialProfile{
-		Kind:      "username_password",
-		Label:     profileLabel,
-		Public:    public,
-		Secret:    map[string]any{"password": password},
-		RiskLabel: provisionRiskLabel(preset),
-		Result: connectors.ActionResult{
-			Status:      connectors.ResultCompleted,
-			Output:      map[string]any{"role_name": roleName, "profile_label": profileLabel, "preset": preset, "scope": summary},
-			DisplayText: "Created Postgres role and saved credential profile",
-		},
-	}, nil
-}
-
-func (Connector) PreserveProvisionedCredentialPublic(existing connectors.CredentialProfileView, requested map[string]any) (map[string]any, error) {
-	if requested == nil {
-		requested = map[string]any{}
-	}
-	if !boolPublic(existing.Public, "managed_by_aipermission") {
-		return clonePublicMap(requested), nil
-	}
-	next := clonePublicMap(requested)
-	existingUsername := strings.TrimSpace(publicString(existing.Public, "username"))
-	requestedUsername := strings.TrimSpace(publicString(next, "username"))
-	if requestedUsername != "" && existingUsername != "" && requestedUsername != existingUsername {
-		return nil, fmt.Errorf("managed credential username cannot be changed; create a new managed profile instead")
-	}
-	if existingUsername != "" {
-		next["username"] = existingUsername
-	}
-	for _, key := range []string{
-		"managed_by_aipermission",
-		"managed_role_name",
-		"managed_admin_profile_id",
-		"managed_admin_profile_ref",
-		"managed_preset",
-		"managed_scope",
-	} {
-		if value, ok := existing.Public[key]; ok {
-			next[key] = value
-		}
-	}
-	return next, nil
-}
-
-func (Connector) ProvisionedCredentialAdminProfileID(profile connectors.CredentialProfileView) (int64, bool, error) {
-	if !boolPublic(profile.Public, "managed_by_aipermission") {
-		return 0, false, nil
-	}
-	adminProfileID := int64Public(profile.Public, "managed_admin_profile_id")
-	if adminProfileID < 1 || adminProfileID == profile.ID {
-		return 0, true, fmt.Errorf("managed credential profile is missing a valid admin profile reference")
-	}
-	return adminProfileID, true, nil
-}
-
-func (Connector) CleanupProvisionedCredentialProfile(ctx context.Context, runtime connectors.RuntimeContext, profile connectors.CredentialProfileView) (connectors.ActionResult, error) {
-	if runtime.Target.ConnectorKind != Kind {
-		return connectors.ActionResult{}, fmt.Errorf("target connector kind must be %s", Kind)
-	}
-	if !boolPublic(profile.Public, "managed_by_aipermission") {
-		return connectors.ActionResult{Status: connectors.ResultCompleted, DisplayText: "No external cleanup required"}, nil
-	}
-	roleName := cleanSimpleIdentifierValue(publicString(profile.Public, "managed_role_name"))
-	if roleName == "" {
-		roleName = cleanSimpleIdentifierValue(publicString(profile.Public, "username"))
-	}
-	if roleName == "" {
-		return connectors.ActionResult{}, fmt.Errorf("managed Postgres profile is missing a role name")
-	}
-	adminRole := cleanSimpleIdentifierValue(publicString(runtime.Profile.Public, "username"))
-	if adminRole == "" {
-		return connectors.ActionResult{}, fmt.Errorf("admin profile is missing a username for managed role cleanup")
-	}
-	ctx, cancel := context.WithTimeout(ctx, queryTimeout)
-	defer cancel()
-	conn, err := connectPostgres(ctx, runtime, false)
-	if err != nil {
-		return connectors.ActionResult{}, err
-	}
-	defer conn.Close(context.Background())
-	var exists bool
-	if err := conn.QueryRow(ctx, `SELECT EXISTS (SELECT 1 FROM pg_roles WHERE rolname = $1)`, roleName).Scan(&exists); err != nil {
-		return connectors.ActionResult{}, fmt.Errorf("check postgres role: %w", err)
-	}
-	if !exists {
-		return connectors.ActionResult{
-			Status:      connectors.ResultCompleted,
-			Output:      map[string]any{"role_name": roleName, "dropped": false, "reason": "role not found"},
-			DisplayText: "Managed Postgres role was already absent",
-		}, nil
-	}
-	tx, err := conn.Begin(ctx)
-	if err != nil {
-		return connectors.ActionResult{}, fmt.Errorf("start postgres role cleanup transaction: %w", err)
-	}
-	defer tx.Rollback(context.Background())
-	for _, statement := range []string{
-		fmt.Sprintf("REASSIGN OWNED BY %s TO %s", quoteIdentifier(roleName), quoteIdentifier(adminRole)),
-		fmt.Sprintf("DROP OWNED BY %s", quoteIdentifier(roleName)),
-		fmt.Sprintf("DROP ROLE %s", quoteIdentifier(roleName)),
-	} {
-		if _, err := tx.Exec(ctx, statement); err != nil {
-			return connectors.ActionResult{}, fmt.Errorf("cleanup postgres role: %w", err)
-		}
-	}
-	if err := tx.Commit(ctx); err != nil {
-		return connectors.ActionResult{}, fmt.Errorf("commit postgres role cleanup: %w", err)
-	}
-	return connectors.ActionResult{
-		Status:      connectors.ResultCompleted,
-		Output:      map[string]any{"role_name": roleName, "dropped": true},
-		DisplayText: "Dropped managed Postgres role",
-	}, nil
-}
-
-func (Connector) Backup(ctx context.Context, runtime connectors.RuntimeContext, _ connectors.BackupRequest) (connectors.BackupArtifact, error) {
-	ctx, cancel := context.WithTimeout(ctx, backupTimeout)
-	defer cancel()
-	invocation, err := postgresCLIConnection(ctx, runtime)
-	if err != nil {
-		return connectors.BackupArtifact{}, err
-	}
-	defer invocation.Cleanup()
-	args := invocation.Args
-	args = append(args,
-		"--format=plain",
-		"--clean",
-		"--if-exists",
-		"--no-owner",
-		"--no-privileges",
-	)
-	var stdout limitedBuffer
-	stdout.Limit = maxBackupBytes
-	var stderr limitedBuffer
-	stderr.Limit = maxRestoreLog
-	cmd := exec.CommandContext(ctx, "pg_dump", args...)
-	cmd.Env = invocation.Env
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	if err := cmd.Run(); err != nil {
-		return connectors.BackupArtifact{}, postgresCommandError("pg_dump", err, stderr.String())
-	}
-	database := targetString(runtime.Target.Config, "database")
-	filename := postgresSafeFilename(runtime.Target.Name, database) + ".sql"
-	return connectors.BackupArtifact{
-		Filename:    filename,
-		ContentType: "application/sql; charset=utf-8",
-		Data:        stdout.Bytes(),
-		Metadata: map[string]any{
-			"connector_kind": Kind,
-			"database":       database,
-			"format":         "plain_sql",
-			"clean":          true,
-		},
-	}, nil
-}
-
-func (Connector) Restore(ctx context.Context, runtime connectors.RuntimeContext, request connectors.RestoreRequest) (connectors.ActionResult, error) {
-	if request.Content == nil || request.Size == 0 {
-		return connectors.ActionResult{}, fmt.Errorf("restore SQL file is empty")
-	}
-	if request.Size < 0 || request.Size > maxBackupBytes {
-		return connectors.ActionResult{}, fmt.Errorf("restore SQL file is too large; maximum restore size is 256 MiB")
-	}
-	ctx, cancel := context.WithTimeout(ctx, restoreTimeout)
-	defer cancel()
-	invocation, err := postgresCLIConnection(ctx, runtime)
-	if err != nil {
-		return connectors.ActionResult{}, err
-	}
-	defer invocation.Cleanup()
-	args := invocation.Args
-	args = append(args,
-		"--single-transaction",
-		"--set", "ON_ERROR_STOP=on",
-	)
-	var stdout limitedBuffer
-	stdout.Limit = maxRestoreLog
-	var stderr limitedBuffer
-	stderr.Limit = maxRestoreLog
-	cmd := exec.CommandContext(ctx, "psql", args...)
-	cmd.Env = invocation.Env
-	cmd.Stdin = io.LimitReader(request.Content, maxBackupBytes+1)
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	if err := cmd.Run(); err != nil {
-		return connectors.ActionResult{}, postgresCommandError("psql", err, stderr.String())
-	}
-	output := map[string]any{
-		"filename": strings.TrimSpace(request.Filename),
-		"stdout":   stdout.String(),
-		"stderr":   stderr.String(),
-	}
-	return connectors.ActionResult{
-		Status:      connectors.ResultCompleted,
-		Output:      output,
-		DisplayText: "Postgres SQL restore completed",
-		Metadata: map[string]any{
-			"connector_kind": Kind,
-			"database":       targetString(runtime.Target.Config, "database"),
-			"filename":       strings.TrimSpace(request.Filename),
-		},
-	}, nil
-}
-
-func connect(ctx context.Context, runtime connectors.RuntimeContext) (*pgx.Conn, error) {
-	return connectPostgres(ctx, runtime, true)
-}
-
-type postgresCLIInvocation struct {
-	Env     []string
-	Args    []string
-	Cleanup func()
-}
-
-func postgresCLIConnection(ctx context.Context, runtime connectors.RuntimeContext) (postgresCLIInvocation, error) {
-	username := strings.TrimSpace(publicString(runtime.Profile.Public, "username"))
-	if username == "" {
-		return postgresCLIInvocation{}, fmt.Errorf("%w: username", ErrMissingSecret)
-	}
-	if runtime.Secrets == nil {
-		return postgresCLIInvocation{}, fmt.Errorf("%w: password", ErrMissingSecret)
-	}
-	password, err := runtime.Secrets.GetSecret(ctx, "password")
-	if err != nil || strings.TrimSpace(password) == "" {
-		return postgresCLIInvocation{}, fmt.Errorf("%w: password", ErrMissingSecret)
-	}
-	host := targetString(runtime.Target.Config, "host")
-	database := targetString(runtime.Target.Config, "database")
-	if host == "" {
-		return postgresCLIInvocation{}, fmt.Errorf("%w: host is required", ErrInvalidConfig)
-	}
-	if database == "" {
-		return postgresCLIInvocation{}, fmt.Errorf("%w: database is required", ErrInvalidConfig)
-	}
-	port := targetPort(runtime.Target.Config)
-	cleanup := func() {}
-	if connectionMode(runtime.Target) == "over_ssh" {
-		localHost, localPort, stop, err := startPostgresTunnel(ctx, runtime)
-		if err != nil {
-			return postgresCLIInvocation{}, err
-		}
-		host = localHost
-		port = localPort
-		cleanup = stop
-	}
-	env := append(os.Environ(),
-		"PGPASSWORD="+password,
-		"PGSSLMODE="+sslMode(runtime.Target.Config),
-		"PGCONNECT_TIMEOUT=10",
-		"PGAPPNAME=aipermission",
-	)
-	args := []string{
-		"--host", host,
-		"--port", strconv.Itoa(port),
-		"--username", username,
-		"--dbname", database,
-		"--no-password",
-	}
-	return postgresCLIInvocation{Env: env, Args: args, Cleanup: cleanup}, nil
-}
-
-func connectPostgres(ctx context.Context, runtime connectors.RuntimeContext, readOnly bool) (*pgx.Conn, error) {
-	username := strings.TrimSpace(publicString(runtime.Profile.Public, "username"))
-	if username == "" {
-		return nil, fmt.Errorf("%w: username", ErrMissingSecret)
-	}
-	if runtime.Secrets == nil {
-		return nil, fmt.Errorf("%w: password", ErrMissingSecret)
-	}
-	password, err := runtime.Secrets.GetSecret(ctx, "password")
-	if err != nil || strings.TrimSpace(password) == "" {
-		return nil, fmt.Errorf("%w: password", ErrMissingSecret)
-	}
-
-	host := targetString(runtime.Target.Config, "host")
-	database := targetString(runtime.Target.Config, "database")
-	if host == "" {
-		return nil, fmt.Errorf("%w: host is required", ErrInvalidConfig)
-	}
-	if database == "" {
-		return nil, fmt.Errorf("%w: database is required", ErrInvalidConfig)
-	}
-
-	connURL := url.URL{
-		Scheme: "postgres",
-		User:   url.UserPassword(username, password),
-		Host:   net.JoinHostPort(host, strconv.Itoa(targetPort(runtime.Target.Config))),
-		Path:   "/" + database,
-	}
-	query := connURL.Query()
-	query.Set("sslmode", sslMode(runtime.Target.Config))
-	query.Set("connect_timeout", "10")
-	connURL.RawQuery = query.Encode()
-
-	config, err := pgx.ParseConfig(connURL.String())
-	if err != nil {
-		return nil, fmt.Errorf("parse postgres connection config: %w", err)
-	}
-	transport, err := postgresNetworkTransport(runtime)
-	if err != nil {
-		return nil, err
-	}
-	dialRequest := postgresNetworkDialRequest(runtime.Target)
-	config.Config.DialFunc = func(ctx context.Context, network string, address string) (net.Conn, error) {
-		return transport.DialConnectorTCP(ctx, dialRequest)
-	}
-
-	conn, err := pgx.ConnectConfig(ctx, config)
-	if err != nil {
-		return nil, fmt.Errorf("connect postgres: %w", err)
-	}
-	if err := configurePostgresSession(ctx, conn); err != nil {
-		_ = conn.Close(ctx)
-		return nil, err
-	}
-	return conn, nil
-}
-
-func configurePostgresSession(ctx context.Context, conn *pgx.Conn) error {
-	if _, err := conn.Exec(ctx, `SELECT set_config('application_name', 'aipermission', false)`); err != nil {
-		return fmt.Errorf("configure postgres application name: %w", err)
-	}
-	statementTimeout := strconv.Itoa(int(queryTimeout.Milliseconds())) + "ms"
-	if _, err := conn.Exec(ctx, `SELECT set_config('statement_timeout', $1, false)`, statementTimeout); err != nil {
-		return fmt.Errorf("configure postgres statement timeout: %w", err)
-	}
-	return nil
-}
-
-func postgresNetworkTransport(runtime connectors.RuntimeContext) (connectors.NetworkTransport, error) {
-	transport, _ := runtime.Capability(connectors.NetworkTransportCapabilityName).(connectors.NetworkTransport)
-	if transport == nil {
-		return nil, ErrMissingTransport
-	}
-	return transport, nil
-}
-
-func postgresNetworkDialRequest(target connectors.TargetView) connectors.NetworkDialRequest {
-	return connectors.NetworkDialRequest{
-		SourceTargetRef:    target.Ref,
-		SourceProjectID:    target.ProjectID,
-		Mode:               connectionMode(target),
-		Host:               targetString(target.Config, "host"),
-		Port:               targetPort(target.Config),
-		TransportTargetRef: strings.TrimSpace(targetString(target.Config, "transport_target_ref")),
-	}
-}
-
-func startPostgresTunnel(ctx context.Context, runtime connectors.RuntimeContext) (string, int, func(), error) {
-	transport, err := postgresNetworkTransport(runtime)
-	if err != nil {
-		return "", 0, nil, err
-	}
-	request := postgresNetworkDialRequest(runtime.Target)
-	listener, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		return "", 0, nil, fmt.Errorf("start postgres local tunnel: %w", err)
-	}
-	tunnelCtx, cancel := context.WithCancel(ctx)
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		for {
-			localConn, err := listener.Accept()
-			if err != nil {
-				return
-			}
-			go pipePostgresTunnelConn(tunnelCtx, transport, request, localConn)
-		}
-	}()
-	addr, ok := listener.Addr().(*net.TCPAddr)
-	if !ok {
-		cancel()
-		_ = listener.Close()
-		<-done
-		return "", 0, nil, fmt.Errorf("postgres local tunnel address is not TCP")
-	}
-	cleanup := func() {
-		cancel()
-		_ = listener.Close()
-		<-done
-	}
-	return "127.0.0.1", addr.Port, cleanup, nil
-}
-
-func pipePostgresTunnelConn(ctx context.Context, transport connectors.NetworkTransport, request connectors.NetworkDialRequest, localConn net.Conn) {
-	remoteConn, err := transport.DialConnectorTCP(ctx, request)
-	if err != nil {
-		_ = localConn.Close()
-		return
-	}
-	copyDone := make(chan struct{}, 2)
-	go func() {
-		_, _ = io.Copy(remoteConn, localConn)
-		_ = remoteConn.Close()
-		_ = localConn.Close()
-		copyDone <- struct{}{}
-	}()
-	go func() {
-		_, _ = io.Copy(localConn, remoteConn)
-		_ = localConn.Close()
-		_ = remoteConn.Close()
-		copyDone <- struct{}{}
-	}()
-	<-copyDone
-}
-
-func provisionScopeInput(input map[string]any) (provisionScope, error) {
-	raw := input["scope"]
-	if raw == nil {
-		return provisionScope{AllSchemas: true}, nil
-	}
-	if text, ok := raw.(string); ok {
-		if strings.TrimSpace(text) == "" {
-			return provisionScope{AllSchemas: true}, nil
-		}
-		var decoded map[string]any
-		if err := json.Unmarshal([]byte(text), &decoded); err != nil {
-			return provisionScope{}, fmt.Errorf("scope must be a JSON object")
-		}
-		raw = decoded
-	}
-	scopeMap, ok := raw.(map[string]any)
-	if !ok {
-		return provisionScope{}, fmt.Errorf("scope must be a JSON object")
-	}
-	scope := provisionScope{AllSchemas: boolInput(scopeMap, "all_schemas")}
-	if scope.AllSchemas {
-		return scope, nil
-	}
-	for _, item := range anySlice(scopeMap["schemas"]) {
-		schemaMap, ok := item.(map[string]any)
-		if !ok {
-			return provisionScope{}, fmt.Errorf("scope schemas must be objects")
-		}
-		schema := provisionSchemaScope{
-			Schema:    cleanSimpleIdentifierValue(stringInput(schemaMap, "schema")),
-			AllTables: boolInput(schemaMap, "all_tables"),
-		}
-		if schema.Schema == "" {
-			return provisionScope{}, fmt.Errorf("scope schema is required and must be a simple identifier")
-		}
-		if !schema.AllTables {
-			for _, tableItem := range anySlice(schemaMap["tables"]) {
-				tableMap, ok := tableItem.(map[string]any)
-				if !ok {
-					return provisionScope{}, fmt.Errorf("scope tables must be objects")
-				}
-				table := provisionTableScope{
-					Table:      cleanSimpleIdentifierValue(stringInput(tableMap, "table")),
-					AllColumns: boolInput(tableMap, "all_columns"),
-				}
-				if table.Table == "" {
-					return provisionScope{}, fmt.Errorf("scope table is required and must be a simple identifier")
-				}
-				if !table.AllColumns {
-					for _, column := range stringSlice(tableMap["columns"]) {
-						clean := cleanSimpleIdentifierValue(column)
-						if clean == "" {
-							return provisionScope{}, fmt.Errorf("scope column is required and must be a simple identifier")
-						}
-						table.Columns = append(table.Columns, clean)
-					}
-					if len(table.Columns) == 0 {
-						return provisionScope{}, fmt.Errorf("selected table must grant all columns or at least one column")
-					}
-				}
-				schema.Tables = append(schema.Tables, table)
-			}
-			if len(schema.Tables) == 0 {
-				return provisionScope{}, fmt.Errorf("selected schema must grant all tables or at least one table")
-			}
-		}
-		scope.Schemas = append(scope.Schemas, schema)
-	}
-	if len(scope.Schemas) == 0 {
-		return provisionScope{}, fmt.Errorf("scope must include at least one schema or all_schemas=true")
-	}
-	return scope, nil
-}
-
-func provisionRoleStatements(target connectors.TargetView, roleName string, password string, preset string, scope provisionScope) ([]string, map[string]any, error) {
-	database := targetString(target.Config, "database")
-	if database == "" {
-		return nil, nil, fmt.Errorf("target database is required")
-	}
-	roleSQL := quoteIdentifier(roleName)
-	statements := []string{fmt.Sprintf("CREATE ROLE %s LOGIN PASSWORD %s", roleSQL, quoteLiteral(password))}
-	statements = append(statements, fmt.Sprintf("GRANT CONNECT ON DATABASE %s TO %s", quoteIdentifier(database), roleSQL))
-	grants := []map[string]any{}
-	privileges := "SELECT"
-	if preset == "read_write" {
-		privileges = "SELECT, INSERT, UPDATE, DELETE"
-	}
-	if scope.AllSchemas {
-		grants = append(grants, map[string]any{"all_schemas": true, "all_tables": true, "privileges": privileges})
-		statements = append(statements, fmt.Sprintf(`
-DO $$
-DECLARE schema_name text;
-BEGIN
-	FOR schema_name IN
-		SELECT nspname FROM pg_namespace
-		WHERE nspname NOT LIKE 'pg_%%' AND nspname <> 'information_schema'
-	LOOP
-		EXECUTE format('GRANT USAGE ON SCHEMA %%I TO %%I', schema_name, %s);
-		EXECUTE format('GRANT %s ON ALL TABLES IN SCHEMA %%I TO %%I', schema_name, %s);
-	END LOOP;
-END
-$$`, quoteLiteral(roleName), privileges, quoteLiteral(roleName)))
-		return statements, map[string]any{"preset": preset, "database": database, "grants": grants}, nil
-	}
-	for _, schema := range scope.Schemas {
-		schemaSQL := quoteIdentifier(schema.Schema)
-		statements = append(statements, fmt.Sprintf("GRANT USAGE ON SCHEMA %s TO %s", schemaSQL, roleSQL))
-		if schema.AllTables {
-			statements = append(statements, fmt.Sprintf("GRANT %s ON ALL TABLES IN SCHEMA %s TO %s", privileges, schemaSQL, roleSQL))
-			grants = append(grants, map[string]any{"schema": schema.Schema, "all_tables": true, "privileges": privileges})
-			continue
-		}
-		for _, table := range schema.Tables {
-			if !table.AllColumns && preset == "read_write" {
-				return nil, nil, fmt.Errorf("column-scoped read/write grants are not supported; choose all columns for write access")
-			}
-			if table.AllColumns {
-				statements = append(statements, fmt.Sprintf("GRANT %s ON TABLE %s TO %s", privileges, qualifiedIdentifierSQL(schema.Schema, table.Table), roleSQL))
-				grants = append(grants, map[string]any{"schema": schema.Schema, "table": table.Table, "all_columns": true, "privileges": privileges})
-				continue
-			}
-			columnSQL := make([]string, 0, len(table.Columns))
-			for _, column := range table.Columns {
-				columnSQL = append(columnSQL, quoteIdentifier(column))
-			}
-			statements = append(statements, fmt.Sprintf("GRANT SELECT (%s) ON TABLE %s TO %s", strings.Join(columnSQL, ", "), qualifiedIdentifierSQL(schema.Schema, table.Table), roleSQL))
-			grants = append(grants, map[string]any{"schema": schema.Schema, "table": table.Table, "columns": table.Columns, "privileges": "SELECT"})
-		}
-	}
-	return statements, map[string]any{"preset": preset, "database": database, "grants": grants}, nil
-}
-
-func randomCredentialPassword() (string, error) {
-	buffer := make([]byte, 32)
-	if _, err := rand.Read(buffer); err != nil {
-		return "", fmt.Errorf("generate password: %w", err)
-	}
-	return base64.RawURLEncoding.EncodeToString(buffer), nil
-}
-
-func provisionRiskLabel(preset string) string {
-	if preset == "read_write" {
-		return "managed read-write"
-	}
-	return "managed read-only"
-}
-
-func boolPublic(public map[string]any, name string) bool {
-	if public == nil {
-		return false
-	}
-	value, ok := public[name]
-	if !ok || value == nil {
-		return false
-	}
-	switch typed := value.(type) {
-	case bool:
-		return typed
-	case string:
-		return strings.EqualFold(strings.TrimSpace(typed), "true")
-	default:
-		return false
-	}
-}
-
-func getTables(ctx context.Context, tx pgx.Tx, schema string, includeSystem bool) (queryOutput, error) {
-	query := `
-		SELECT table_schema, table_name, table_type
-		FROM information_schema.tables
-		WHERE ($1 = '' OR table_schema = $1)`
-	if !includeSystem {
-		query += ` AND table_schema NOT IN ('pg_catalog', 'information_schema') AND table_schema NOT LIKE 'pg_toast%'`
-	}
-	query += ` ORDER BY table_schema, table_name`
-	return queryRows(ctx, tx, query, 1000, schema)
-}
-
-func describeTable(ctx context.Context, tx pgx.Tx, schema string, table string) (queryOutput, error) {
-	query := `
-		SELECT table_schema, table_name, ordinal_position, column_name, data_type, is_nullable, column_default
-		FROM information_schema.columns
-		WHERE table_name = $1
-			AND ($2 = '' OR table_schema = $2)
-			AND table_schema NOT IN ('pg_catalog', 'information_schema')
-		ORDER BY table_schema, table_name, ordinal_position`
-	return queryRows(ctx, tx, query, 500, table, schema)
-}
-
-func queryRows(ctx context.Context, tx pgx.Tx, sql string, rowLimit int, args ...any) (queryOutput, error) {
-	if rowLimit < 1 {
-		rowLimit = defaultMaxRows
-	}
-	if rowLimit > maxRows {
-		rowLimit = maxRows
-	}
-	rows, err := tx.Query(ctx, sql, args...)
-	if err != nil {
-		return queryOutput{}, fmt.Errorf("query postgres: %w", err)
-	}
-	defer rows.Close()
-
-	fields := rows.FieldDescriptions()
-	columns := make([]string, 0, len(fields))
-	for _, field := range fields {
-		columns = append(columns, field.Name)
-	}
-	builder := sqlresult.NewBuilder(columns, rowLimit, maxOutputBytes, maxCellBytes, truncatedSuffix)
-	for rows.Next() {
-		values, err := rows.Values()
-		if err != nil {
-			return queryOutput{}, fmt.Errorf("read postgres row: %w", err)
-		}
-		if !builder.Add(values, normalizePostgresValue) {
-			break
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return queryOutput{}, fmt.Errorf("iterate postgres rows: %w", err)
-	}
-	return queryOutput{Result: builder.Result(nil)}, nil
-}
-
-func normalizePostgresValue(value any) any {
-	switch typed := value.(type) {
-	case []byte:
-		return string(typed)
-	case time.Time:
-		return typed.UTC().Format(time.RFC3339Nano)
-	default:
-		return typed
-	}
-}
-
-func truncateUTF8Bytes(value string, limit int) string {
-	if limit <= 0 {
-		return ""
-	}
-	if len(value) <= limit {
-		return value
-	}
-	cut := 0
-	for index := range value {
-		if index > limit {
-			break
-		}
-		cut = index
-	}
-	if cut == 0 {
-		return ""
-	}
-	return value[:cut]
-}
-
-type limitedBuffer struct {
-	Limit int
-	data  bytes.Buffer
-}
-
-func (b *limitedBuffer) Write(p []byte) (int, error) {
-	if b.Limit <= 0 {
-		return len(p), nil
-	}
-	remaining := b.Limit - b.data.Len()
-	if remaining <= 0 {
-		return 0, fmt.Errorf("postgres command output exceeded %d bytes", b.Limit)
-	}
-	if len(p) > remaining {
-		_, _ = b.data.Write(p[:remaining])
-		return remaining, fmt.Errorf("postgres command output exceeded %d bytes", b.Limit)
-	}
-	return b.data.Write(p)
-}
-
-func (b *limitedBuffer) Bytes() []byte {
-	return b.data.Bytes()
-}
-
-func (b *limitedBuffer) String() string {
-	return b.data.String()
-}
-
-func postgresCommandError(command string, err error, stderr string) error {
-	var execErr *exec.Error
-	if errors.As(err, &execErr) {
-		return fmt.Errorf("%s is not available in the gateway container; rebuild with postgresql-client installed", command)
-	}
-	message := strings.TrimSpace(truncateUTF8Bytes(stderr, 4000))
-	if message == "" {
-		message = err.Error()
-	}
-	if errors.Is(err, io.ErrShortWrite) {
-		message = "command output exceeded gateway limit"
-	}
-	if command == "pg_dump" && strings.Contains(message, "server version mismatch") {
-		message += "\nThe gateway pg_dump client is older than this Postgres server. Rebuild the AIPermission backend image so the bundled Postgres client is updated."
-	}
-	return fmt.Errorf("%s failed: %s", command, message)
-}
-
-func postgresSafeFilename(parts ...string) string {
-	candidate := strings.Join(parts, "-")
-	candidate = strings.ToLower(strings.TrimSpace(candidate))
-	candidate = regexp.MustCompile(`[^a-z0-9._-]+`).ReplaceAllString(candidate, "-")
-	candidate = strings.Trim(candidate, "-._")
-	if candidate == "" {
-		return "postgres-backup"
-	}
-	return candidate
 }
 
 func connectionMode(target connectors.TargetView) string {
